+++ src/components/Code/index.js
@@ -9,9 +9,6 @@ const HighlightedCode = (code, isHighlighted, language, className, testSection) 
   if (isHighlighted) {
     // Code that uses syntax highlighting needs to have
     // `dangerouslySetInnerHTML` set so that the HTML returned is displayed.
-    // use require ! prefix to ignore local webpack config
-    require('!style-loader!css-loader!highlight.js/styles/docco.css');
-
     dangerouslySetInnerHTML = {
       __html: language ? Highlight.highlight(language, code).value :
                          Highlight.highlightAuto(code).value,