+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## [Unreleased][Unreleased]
+
+## [17.3.2][17.3.2] - 2016-09-19
 ### Changed
 - [Patch] Remove `css-loader` dependency that was used by highlight.js and implement it the old-fashioned way. (#579)
 
@@ -573,7 +575,7 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 ### Added
 - [Patch] Use semantic versioning in LEGO (#58).
 
-[Unreleased]: https://github.com/optimizely/oui/compare/v17.3.1...HEAD
+[Unreleased]: https://github.com/optimizely/oui/compare/v17.3.2...HEAD
 [0.0.2]: https://github.com/optimizely/oui/compare/v0.0.1...v0.0.2
 [0.0.3]: https://github.com/optimizely/oui/compare/v0.0.2...v0.0.3
 [1.0.0]: https://github.com/optimizely/oui/compare/v0.0.3...v1.0.0
@@ -628,3 +630,4 @@ This file follows the format suggested by [Keep a CHANGELOG](https://github.com/
 [17.2.0]: https://github.com/optimizely/oui/compare/v17.1.0...v17.2.0
 [17.3.0]: https://github.com/optimizely/oui/compare/v17.2.0...v17.3.0
 [17.3.1]: https://github.com/optimizely/oui/compare/v17.3.0...v17.3.1
+[17.3.2]: https://github.com/optimizely/oui/compare/v17.3.1...v17.3.2